db table bugfix

+++ share/db_installer/sql/tags_settings.sql.php
@@ -1,6 +1,6 @@
 <?php
 return '
-  `user_id` int() unsigned NOT NULL,
-  `allowed_group` tinyint() unsigned NOT NULL,
+  `user_id` int(11) unsigned NOT NULL,
+  `allowed_group` tinyint(3) unsigned NOT NULL,
   PRIMARY KEY (`user_id`)
 ';
